requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
alembic==1.13.2
paho-mqtt==2.1.0
pandas==2.2.3
APScheduler==3.10.4
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.3
pyarrow==14.0.1
requests==2.31.0